# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/competitions/current": {
      "get": {
        "operationId": "getCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Return the authenticated user current regional competition",
        "tags": ["competitions"]
      }
    },
    "/v1/competitions/{competitionId}/enrollments": {
      "post": {
        "operationId": "enroll",
        "parameters": [
          {
            "name": "competitionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEnrollmentDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Enroll the authenticated eligible user in a goal bracket",
        "tags": ["competitions"]
      }
    },
    "/v1/competitions/{monthKey}/enrollment-count": {
      "get": {
        "operationId": "getEnrollmentCount",
        "parameters": [
          {
            "name": "monthKey",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentCountResponseDto"
                }
              }
            }
          }
        },
        "summary": "Return the active enrollment count for a regional competition",
        "tags": ["competitions"]
      }
    },
    "/v1/competitions/{monthKey}/matches": {
      "get": {
        "operationId": "getMatches",
        "parameters": [
          {
            "name": "monthKey",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "goal",
            "required": true,
            "in": "query",
            "schema": { "minimum": 1, "maximum": 7, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetitionMatchResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Return the authenticated entrant weekly competition matches",
        "tags": ["competitions"]
      }
    },
    "/v1/legal-documents/current": {
      "get": {
        "operationId": "getCurrent",
        "parameters": [
          {
            "name": "jurisdictionCode",
            "required": false,
            "in": "query",
            "schema": { "default": "GLOBAL", "type": "string" }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": { "default": "en", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentLegalDocumentsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Return the current server-authoritative legal document bundle for a jurisdiction and locale",
        "tags": ["legal-documents"]
      }
    },
    "/v1/me/legal-receipts/status": {
      "get": {
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "jurisdictionCode",
            "required": false,
            "in": "query",
            "schema": { "default": "GLOBAL", "type": "string" }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": { "default": "en", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalReceiptStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Return this account current legal receipt status",
        "tags": ["legal-receipts"]
      }
    },
    "/v1/me/legal-receipts": {
      "post": {
        "operationId": "record",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordLegalReceiptBundleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalReceiptStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Atomically record receipts for the exact current legal bundle",
        "tags": ["legal-receipts"]
      }
    },
    "/v1/operator/configuration/legal-documents": {
      "post": {
        "operationId": "publish",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishLegalDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLegalDocumentResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Publish an immutable legal document version",
        "tags": ["operator-configuration"]
      }
    },
    "/v1/operator/configuration/legal-documents/{legalDocumentId}/withdrawal": {
      "post": {
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "legalDocumentId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawLegalDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLegalDocumentResponseDto"
                }
              }
            }
          }
        },
        "security": [{ "firebase": [] }],
        "summary": "Withdraw an immutable legal document version",
        "tags": ["operator-configuration"]
      }
    },
...
  },
  "info": {
    "title": "GoGymGo API",
    "description": "Authoritative API for GoGymGo accounts, verified competition activity, ledgers, results, and payouts.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "firebase": {
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "type": "http",
        "description": "Firebase ID token verified by the GoGymGo API."
      },
      "hyperwallet-webhook": {
        "type": "http",
        "scheme": "basic",
        "description": "Dedicated credentials configured only for Hyperwallet."
      }
    },
    "schemas": {
...
      "CompetitionRulesResponseDto": {
        "type": "object",
        "properties": {
          "minSessionMinutes": { "type": "number" },
          "minHeartRateSamples": { "type": "number" },
          "requireDeviceAttestation": { "type": "boolean" },
          "requireFaceCheck": { "type": "boolean" },
          "requireGymQr": { "type": "boolean" },
          "signupPrizeDrawEntries": { "type": "number" },
          "verifiedSessionCategoryScore": { "type": "number" },
          "verifiedSessionPrizeDrawEntries": { "type": "number" },
          "payoutPoolAmountMinor": { "type": "number" },
          "payoutWinnerCount": { "type": "number" },
          "payoutExponent": { "type": "number" }
        },
        "required": [
          "minSessionMinutes",
          "minHeartRateSamples",
          "requireDeviceAttestation",
          "requireFaceCheck",
          "requireGymQr",
          "signupPrizeDrawEntries",
          "verifiedSessionCategoryScore",
          "verifiedSessionPrizeDrawEntries",
          "payoutPoolAmountMinor",
          "payoutWinnerCount",
          "payoutExponent"
        ]
      },
      "CompetitionResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "monthKey": { "type": "string" },
          "name": { "type": "string" },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "registration",
              "active",
              "settling",
              "settled",
              "cancelled"
            ]
          },
          "currency": { "type": "string" },
          "regionCode": { "type": "string" },
          "regionName": { "type": "string" },
          "rulesVersion": { "type": "string" },
          "rules": {
            "$ref": "#/components/schemas/CompetitionRulesResponseDto"
          },
          "goalDays": { "type": "array", "items": { "type": "number" } },
          "registrationOpensAt": { "type": "string", "format": "date-time" },
          "registrationClosesAt": { "type": "string", "format": "date-time" },
          "startsAt": { "type": "string", "format": "date-time" },
          "endsAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "monthKey",
          "name",
          "status",
          "currency",
          "regionCode",
          "regionName",
          "rulesVersion",
          "rules",
          "goalDays",
          "registrationOpensAt",
          "registrationClosesAt",
          "startsAt",
          "endsAt"
        ]
      },
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {
          "goalDays": { "type": "number", "maximum": 7, "minimum": 1 },
          "regionVerificationId": { "type": "string", "format": "uuid" },
          "legalReceiptBundleId": { "type": "string", "format": "uuid" },
          "rulesAccepted": { "type": "boolean", "enum": [true] },
          "ageEligibilityAttested": { "type": "boolean", "enum": [true] }
        },
        "required": [
          "goalDays",
          "regionVerificationId",
          "legalReceiptBundleId",
          "rulesAccepted",
          "ageEligibilityAttested"
        ]
      },
      "EnrollmentResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "competitionId": { "type": "string", "format": "uuid" },
          "goalDays": { "type": "number" },
          "status": {
            "type": "string",
            "enum": ["active", "withdrawn", "disqualified"]
          },
          "enrolledAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "competitionId", "goalDays", "status", "enrolledAt"]
      },
      "EnrollmentCountResponseDto": {
        "type": "object",
        "properties": { "count": { "type": "number" } },
        "required": ["count"]
      },
      "CompetitionMatchResponseDto": {
        "type": "object",
        "properties": {
          "availability": {
            "type": "string",
            "enum": ["matched", "searching", "solo"]
          },
          "opponentAlias": { "type": "string" },
          "opponentVerifiedDateKeys": {
            "type": "array",
            "items": { "type": "string" }
          },
          "periodIndex": { "type": "number", "maximum": 4, "minimum": 1 },
          "region": { "type": "string" }
        },
        "required": [
          "availability",
          "opponentAlias",
          "opponentVerifiedDateKeys",
          "periodIndex",
          "region"
        ]
      },
      "LegalDocumentSectionDto": {
        "type": "object",
        "properties": {
          "heading": { "type": "string", "maxLength": 160 },
          "body": { "type": "string", "maxLength": 4000 },
          "bullets": {
            "maxItems": 30,
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["heading"]
      },
      "LegalDocumentContentDto": {
        "type": "object",
        "properties": {
          "intro": { "type": "string", "maxLength": 4000 },
          "sections": {
            "maxItems": 40,
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegalDocumentSectionDto" }
          }
        },
        "required": ["intro", "sections"]
      },
      "LegalDocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "documentKey": { "type": "string" },
          "jurisdictionCode": { "type": "string" },
          "locale": { "type": "string" },
          "version": { "type": "string" },
          "title": { "type": "string" },
          "content": { "$ref": "#/components/schemas/LegalDocumentContentDto" },
          "contentSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "receiptRequirement": {
            "type": "string",
            "enum": ["accept", "acknowledge", "none"]
          },
          "effectiveAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "documentKey",
          "jurisdictionCode",
          "locale",
          "version",
          "title",
          "content",
          "contentSha256",
          "receiptRequirement",
          "effectiveAt"
        ]
      },
      "CurrentLegalDocumentsResponseDto": {
        "type": "object",
        "properties": {
          "bundleSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "configured": { "type": "boolean" },
          "documents": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegalDocumentResponseDto" }
          },
          "jurisdictionCode": { "type": "string" },
          "locale": { "type": "string" }
        },
        "required": [
          "bundleSha256",
          "configured",
          "documents",
          "jurisdictionCode",
          "locale"
        ]
      },
      "LegalReceiptStatusResponseDto": {
        "type": "object",
        "properties": {
          "bundleSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "configured": { "type": "boolean" },
          "documents": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegalDocumentResponseDto" }
          },
          "jurisdictionCode": { "type": "string" },
          "locale": { "type": "string" },
          "complete": { "type": "boolean" },
          "receiptBundleId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "bundleSha256",
          "configured",
          "documents",
          "jurisdictionCode",
          "locale",
          "complete"
        ]
      },
      "LegalReceiptItemDto": {
        "type": "object",
        "properties": {
          "documentId": { "type": "string", "format": "uuid" },
          "contentSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "action": { "type": "string", "enum": ["accept", "acknowledge"] }
        },
        "required": ["documentId", "contentSha256", "action"]
      },
      "RecordLegalReceiptBundleDto": {
        "type": "object",
        "properties": {
          "jurisdictionCode": { "type": "string", "default": "GLOBAL" },
          "locale": { "type": "string", "default": "en" },
          "bundleSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "documents": {
            "maxItems": 20,
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegalReceiptItemDto" }
          }
        },
        "required": ["bundleSha256", "documents"]
      },
      "PublishLegalDocumentDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 },
          "documentKey": {
            "type": "string",
            "example": "terms_of_service",
            "maxLength": 64
          },
          "jurisdictionCode": {
            "type": "string",
            "example": "CA-BC",
            "maxLength": 16
          },
          "locale": { "type": "string", "example": "en-CA", "maxLength": 16 },
          "version": { "type": "string", "maxLength": 64 },
          "title": { "type": "string", "maxLength": 160 },
          "content": { "$ref": "#/components/schemas/LegalDocumentContentDto" },
          "receiptRequirement": {
            "type": "string",
            "enum": ["accept", "acknowledge", "none"]
          },
          "effectiveAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "reason",
          "documentKey",
          "jurisdictionCode",
          "locale",
          "version",
          "title",
          "content",
          "receiptRequirement",
          "effectiveAt"
        ]
      },
      "AdminLegalDocumentResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["effective", "scheduled", "withdrawn"]
          },
          "contentSha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          }
        },
        "required": ["id", "status", "contentSha256"]
      },
      "WithdrawLegalDocumentDto": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "maxLength": 500, "minLength": 8 }
        },
        "required": ["reason"]
      },
...
    }
  }
}